+++ src/AVPRIndex/Globals.fs
@@ -2,5 +2,4 @@
 
 module Globals =
 
-    let [<Literal>] STAGING_AREA_RELATIVE_PATH = "src/PackageRegistryService/StagingArea"
-    let [<Literal>] PACKAGE_INDEX_RELATIVE_PATH = "src/PackageRegistryService/Data/arc-validate-package-index.json"
+    let [<Literal>] STAGING_AREA_RELATIVE_PATH = "StagingArea"