+++ tests/IndexTests/ValidationPackageIndexTests.fs
@@ -1,0 +1,16 @@
+﻿namespace ValidationPackageIndexTests
+
+open System
+open Xunit
+
+module InMemory =
+
+    [<Fact>]
+    let ``My test`` () =
+        Assert.True(true)
+
+module IO =
+
+    [<Fact>]
+    let ``My test`` () =
+        Assert.True(true)